refactor(bots): rename simple-bridge to telegram-bridge, drop dead feishu/discord branches
The registry only ever constructs this bridge for telegram — feishu
and discord route to their dedicated bridges — so the feishu start
branch (plus feishuTenantAccessToken and its timeout constant), the
discord scaffold branch, and the unreachable 'unimplemented' fallback
were dead. Rename the file and class to match the real responsibility
and simplify connectionKind to a constant 'polling'.

Generated-by: ZCode (GLM)
